+++ packages/expo-router/package.json
@@ -32,7 +32,7 @@
     "@react-navigation/native": "^6.1.7",
     "@react-navigation/native-stack": "^6.9.13",
     "@react-navigation/routers": "^6.1.9",
-    "@tamagui/web": "latest",
+    "@tamagui/web": "1.94.4",
     "escape-string-regexp": "^5.0.0",
     "expo": "^47.0.0",
     "expo-constants": "^14.4.2",
@@ -52,7 +52,7 @@
     "react-native": "*"
   },
   "devDependencies": {
-    "@tamagui/build": "latest",
+    "@tamagui/build": "1.94.4",
     "react": "^18.2.0",
     "react-native": "0.73.2"
   },
