+++ packages/react-native-prebuilt/types/index.d.ts
@@ -1,0 +1,6 @@
+import { type BuildOptions } from 'esbuild';
+export declare function buildAll(): Promise<void>;
+export declare function buildReactJSX(options?: BuildOptions): Promise<void>;
+export declare function buildReact(options?: BuildOptions): Promise<void>;
+export declare function buildReactNative(options?: BuildOptions): Promise<void>;
+//# sourceMappingURL=index.d.ts.map
