+++ packages/vite-native-client/package.json
@@ -27,7 +27,7 @@
     }
   },
   "devDependencies": {
-    "@tamagui/build": "latest",
+    "@tamagui/build": "1.94.4",
     "react-native": "0.73.2"
   },
   "peerDependencies": {
