+++ packages/vite-native-swc/package.json
@@ -29,7 +29,7 @@
     "@swc/core": "^1.3.78"
   },
   "devDependencies": {
-    "@tamagui/build": "latest",
+    "@tamagui/build": "1.94.4",
     "react-native": "0.73.2"
   },
   "peerDependencies": {
