+++ tsconfig.json
@@ -22,8 +22,9 @@
     "inlineSourceMap": true,
     "preserveSymlinks": true,
     "jsx": "react-jsx",
-    "module": "system",
+    "module": "ESNext",
     "moduleResolution": "node",
+    "noEmit": false,
     "noEmitOnError": false,
     "noImplicitAny": false,
     "noImplicitReturns": false,
